+++ R/group_tt.R
@@ -182,7 +182,6 @@ style_group_j <- function(x) {
         j = group_cols,
         line = "b",
         line_width = 0.03,
-        line_color = "black",
         line_trim = line_trim_spec
       )
     }